Create LICENSE
As it stands, the project has no license, which means that it falls under "implied copyright", and that means that nobody can use your code in any other projects, as well as change it to add improvements and submit those as pull requests. (More info on that here)[https://choosealicense.com/no-permission/]. I'm proposing the MIT license because it is short, permissive, and to the point, it also means that those who use your work are still required to credit you as their source.